Move modified code to new CompositeQuestion class
