+++ src/components/Question.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Accordion, Card } from 'react-bootstrap';
+import { Card, Accordion } from 'react-bootstrap';
 import JsonLdUtils from 'jsonld-utils';
 import PropTypes from 'prop-types';
 import Answer from './Answer';
@@ -12,7 +12,7 @@ import ValidatorFactory from '../model/ValidatorFactory';
 import JsonLdObjectUtils from '../util/JsonLdObjectUtils';
 import PrefixIcon from './PrefixIcon';
 import MediaContent from './MediaContent';
-import { CaretSquareDown, CaretSquareUp, InfoCircle } from '../styles/icons';
+import { CaretSquareUp, CaretSquareDown, InfoCircle } from '../styles/icons';
 import { ConfigurationContext } from '../contexts/ConfigurationContext';
 import classNames from 'classnames';
 
@@ -35,92 +35,11 @@ export default class Question extends React.Component {
 
   onAnswerChange = (answerIndex, change) => {
     this._onChange(Constants.HAS_ANSWER, answerIndex, change);
-
-    const question = this.props.question;
-    if (JsonLdUtils.getJsonAttValue(question, Constants.COMPOSITE_PATTERN)) {
-      this._updateExpandedParts(change);
-    }
   };
 
   onSubQuestionChange = (subQuestionIndex, change) => {
     this._onChange(Constants.HAS_SUBQUESTION, subQuestionIndex, change);
-
-    const question = this.props.question;
-    if (JsonLdUtils.getJsonAttValue(question, Constants.COMPOSITE_PATTERN)) {
-      this._updateCollapsedComposite();
-    }
   };
-
-  _updateExpandedParts() {
-    const question = this.props.question;
-    const subQuestions = this._getSubQuestions();
-
-    const patternVariables = question[Constants.COMPOSITE_VARIABLES];
-
-    let regex = new RegExp(question[Constants.PATTERN]);
-    /**
-     * @var {string} value
-     */
-    let value = question[Constants.HAS_ANSWER][0][Constants.HAS_DATA_VALUE]['@value'];
-    let match = value.match(regex);
-
-    if (match) {
-
-      for (let i = 1; i < match.length; i++) {
-        let subValue = match[i];
-        if (!subValue) {
-          subValue = '';
-        }
-
-        let id = patternVariables[i - 1];
-        let subQuestion = this._findById(subQuestions, id);
-
-        subQuestion[Constants.HAS_ANSWER][0][Constants.HAS_DATA_VALUE] = {
-          '@value': subValue.trim()
-        }
-      }
-
-    } else {
-      subQuestions.forEach(subQuestion => {
-        subQuestion[Constants.HAS_ANSWER][0][Constants.HAS_DATA_VALUE] = {
-          '@value': ''
-        }
-      });
-    }
-
-  }
-
-  _findById(arr, id) {
-    return arr.find(el => el['@id'] === id);
-  }
-
-  _updateCollapsedComposite() {
-    const question = this.props.question;
-    const subQuestions = this._getSubQuestions();
-
-    /**
-     * @var {string} compositePattern
-     */
-    let compositePattern = question[Constants.COMPOSITE_PATTERN];
-    const patternVariables = question[Constants.COMPOSITE_VARIABLES];
-
-    for (let i = patternVariables.length; i >= 1; i--) {
-
-      let id = patternVariables[i - 1];
-      let subQuestion = this._findById(subQuestions, id);
-
-      let val = subQuestion[Constants.HAS_ANSWER][0][Constants.HAS_DATA_VALUE]['@value'];
-      if (!val) {
-        val = '';
-      }
-
-      compositePattern = compositePattern.replace('?' + i, val.trim());
-    }
-
-    let change = {};
-    change[Constants.HAS_DATA_VALUE] = {'@value': compositePattern.trim()}
-    this._onChange(Constants.HAS_ANSWER, 0, change);
-  }
 
   _onChange(att, valueIndex, newValue) {
     let newState = { ...this.props.question };
@@ -165,20 +84,6 @@ export default class Question extends React.Component {
         );
       }
     }
-
-    if (JsonLdUtils.getJsonAttValue(question, Constants.COMPOSITE_PATTERN)) {
-      return (
-        <Card className="mb-3">
-          <div className="p-3">
-            {this.renderAnswers()}
-            <div className="p-3">
-              {this.renderSubQuestions()}
-            </div>
-          </div>
-        </Card>
-      );
-    }
-
     if (FormUtils.isSection(question)) {
       const { collapsible, withoutCard } = this.props;
       const categoryClass = Question._getQuestionCategoryClass(question);
@@ -241,8 +146,6 @@ export default class Question extends React.Component {
         FormUtils.isSparqlInput(question) ||
         FormUtils.isTurtleInput(question);
       cls = classNames(Question._getAnswerClass(isTextarea), Question._getQuestionCategoryClass(question));
-
-
       row.push(
         <div key={'row-item-' + i} className={cls} id={question['@id']}>
           <div className="row">
